requirements.txt
streamlit
openai
llama-index
PyMuPDF
extract-msg
pandas
python-docx
beautifulsoup4
